refactor: extract ui keybind help
